pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-tools-cli"
version = "0.3.0"
description = "Python CLI tools for photography workflows"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.24,<1.0",
    "pillow>=12.1,<13.0",
    "python-dotenv>=1.2,<2.0",
]

[dependency-groups]
dev = [
    "pytest>=9,<10",
    "ruff>=0.15,<1.0",
    "mypy>=1.19,<2.0",
    "build>=1.4,<2.0",
    "twine>=6.2,<7.0",
]

[project.scripts]
photo-tools = "photo_tools.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = "src"
explicit_package_bases = true
